+++ packages/bpk-component-spinner/src/BpkExtraLargeSpinner-test.js
@@ -19,11 +19,22 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
+import SPINNER_TYPES from './spinnerTypes';
 import BpkExtraLargeSpinner from './BpkExtraLargeSpinner';
 
 describe('BpkExtraLargeSpinner', () => {
   it('should render correctly', () => {
     const tree = renderer.create(<BpkExtraLargeSpinner />).toJSON();
+    expect(tree).toMatchSnapshot();
+  });
+
+  it('should render correctly with type as light', () => {
+    const tree = renderer.create(<BpkExtraLargeSpinner type={SPINNER_TYPES.light} />).toJSON();
+    expect(tree).toMatchSnapshot();
+  });
+
+  it('should render correctly with type as dark', () => {
+    const tree = renderer.create(<BpkExtraLargeSpinner type={SPINNER_TYPES.dark} />).toJSON();
     expect(tree).toMatchSnapshot();
   });