+++ changelog.md
@@ -2,7 +2,9 @@
 
 ## UNRELEASED
 
-_Nothing yet..._
+**Added:**
+- bpk-component-spinner:
+  - New `type` prop. Options are `primary`, `light` or `dark`
 
 ## 2017-11-21 - React Native cards can now have a dividing line